common file for common defs
