+++ lib/dbservice_web/swagger_schemas/batch.ex
@@ -30,34 +30,4 @@ defmodule DbserviceWeb.SwaggerSchema.Batch do
         end,
     }
   end
-
-  def user_ids do
-    %{
-      UserIds:
-        swagger_schema do
-          properties do
-            user_ids(:array, "List of user ids")
-          end
-
-          example(%{
-            user_ids: [1, 2]
-          })
-        end,
-    }
-  end
-
-  def session_ids do
-    %{
-      SessionIds:
-        swagger_schema do
-          properties do
-            session_ids(:array, "List of session ids")
-          end
-
-          example(%{
-            session_ids: [1, 2]
-          })
-        end
-    }
-  end
 end
