+++ lib/dbservice_web/swagger_schemas/session.ex
@@ -56,19 +56,4 @@ defmodule DbserviceWeb.SwaggerSchema.Session do
         end,
     }
   end
-
-  def batch_ids do
-    %{
-      BatchIds:
-        swagger_schema do
-          properties do
-            batch_ids(:array, "List of batch ids")
-          end
-
-          example(%{
-            batch_ids: [1, 2]
-          })
-        end
-    }
-  end
 end
